refact: refactor the main content compose container
