+++ app/src/main/java/com/lucasmarciano/composeproject/features/home/HomeScreen.kt
@@ -1,18 +1,11 @@
 package com.lucasmarciano.composeproject.features.home
 
 import android.widget.Toast
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.rememberScrollState
-import androidx.compose.foundation.verticalScroll
-import androidx.compose.material.MaterialTheme
 import androidx.compose.runtime.Composable
-import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.LocalContext
 import com.lucasmarciano.composeproject.features.home.components.ShimmerHomeController
 import com.lucasmarciano.composeproject.ui.InterfaceFactory
-import com.lucasmarciano.composeproject.ui.utils.spacing
+import com.lucasmarciano.composeproject.ui.components.MainContainer
 import com.lucasmarciano.composeproject.ui.values.InterfaceItemVO
 
 @Composable
@@ -33,19 +26,8 @@ fun HomeContent(state: HomeUIState) {
 
 @Composable
 private fun MainContent(isLoading: Boolean = true, listItems: List<InterfaceItemVO> = emptyList()) {
-    val scrollState = rememberScrollState()
-
     ShimmerHomeController(isLoading) {
-        Column(
-            modifier = Modifier
-                .fillMaxSize()
-                .padding(
-                    top = MaterialTheme.spacing.medium,
-                    start = MaterialTheme.spacing.medium,
-                    end = MaterialTheme.spacing.medium
-                )
-                .verticalScroll(scrollState)
-        ) {
+        MainContainer {
             InterfaceFactory(listItems)
         }
     }
