+++ app/src/main/java/com/lucasmarciano/composeproject/features/settings/SettingsScreen.kt
@@ -1,18 +1,11 @@
 package com.lucasmarciano.composeproject.features.settings
 
 import android.widget.Toast
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.rememberScrollState
-import androidx.compose.foundation.verticalScroll
-import androidx.compose.material.MaterialTheme
 import androidx.compose.runtime.Composable
-import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.LocalContext
 import com.lucasmarciano.composeproject.features.settings.components.ShimmerSettingsController
 import com.lucasmarciano.composeproject.ui.InterfaceFactory
-import com.lucasmarciano.composeproject.ui.utils.spacing
+import com.lucasmarciano.composeproject.ui.components.MainContainer
 import com.lucasmarciano.composeproject.ui.values.InterfaceItemVO
 
 @Composable
@@ -39,19 +32,8 @@ private fun SettingsScreen(
     isLoading: Boolean = true,
     listItems: List<InterfaceItemVO> = emptyList()
 ) {
-    val scrollState = rememberScrollState()
-
     ShimmerSettingsController(isLoading) {
-        Column(
-            modifier = Modifier
-                .fillMaxSize()
-                .padding(
-                    top = MaterialTheme.spacing.medium,
-                    start = MaterialTheme.spacing.medium,
-                    end = MaterialTheme.spacing.medium
-                )
-                .verticalScroll(scrollState)
-        ) {
+        MainContainer {
             InterfaceFactory(listItems)
         }
     }
